Refactor to replace CcTuningChanger with PedalTuningChanger.
